comment like counter

+++ src/main/java/com/laoumri/supsharespringbootsocialmediaapp/dto/responses/CommentResponse.java
@@ -21,6 +21,6 @@ public class CommentResponse {
     private Instant created_at;
     private Instant updated_at;
     private boolean updated;
-    private int likes_count;
-    private int dislikes_count;
+    private long likes_count;
+    private long dislikes_count;
 }
